Change push default to 'simple'.
The 'matching' option is the default behavior in Git 1.x. It means that
if you do a git push without specifying a branch, it will push all your
local branches to their matching ones on your remote repository.

The new default in Git 2.x is 'simple'. It means that when doing a git
push without specifying a branch, only your current branch will be
pushed to the one git pull would normally get your code from.